- moving all sidebar cards to new folder - utilizing the `isActive` prop of the SibarMenuButtons
