+++ src/components/common/sidebar_buttons/FriendCard.tsx
